Split out workspaceUtils
